# schemas/s3-api.openapi.yaml
openapi: 3.1.0
info:
  title: BleepStore S3-Compatible API
  version: 1.0.0
  description: |
    BleepStore implements a core subset of the Amazon S3 REST API, providing
    S3-compatible object storage via path-style URLs. This schema covers
    Phase 1 operations including bucket management, object CRUD, access-control
    lists, and multipart uploads.

    All XML request and response bodies use the S3 namespace:
    http://s3.amazonaws.com/doc/2006-03-01/
  license:
    name: MIT
servers:
  - url: https://localhost:9000
    description: Local BleepStore development server
  - url: https://s3.bleepstore.example.com
    description: BleepStore production endpoint
security:
  - sigv4: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /:
    get:
      operationId: ListBuckets
      summary: List all buckets owned by the authenticated sender
      tags: [Bucket]
      responses:
        "200":
          description: Bucket list returned successfully.
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/ListAllMyBucketsResult"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # =======================================================================
  # Bucket-level: /{Bucket}
  # =======================================================================
  /{Bucket}:
    parameters:
      - $ref: "#/components/parameters/BucketParam"

    # ----- PUT /{Bucket} --------------------------------------------------
    put:
      operationId: CreateBucket
      summary: Create a new bucket
      tags: [Bucket]
      parameters:
        - $ref: "#/components/parameters/AclQueryParam"
        - name: x-amz-acl
          in: header
          description: Canned ACL to apply to the bucket.
          schema:
            $ref: "#/components/schemas/CannedAcl"
      requestBody:
        description: |
          Optional CreateBucketConfiguration specifying the location constraint.
          When the `acl` query parameter is present this endpoint acts as
          PutBucketAcl and expects an AccessControlPolicy body instead.
        required: false
        content:
          application/xml:
            schema:
              oneOf:
                - $ref: "#/components/schemas/CreateBucketConfiguration"
                - $ref: "#/components/schemas/AccessControlPolicy"
      responses:
        "200":
          description: Bucket created (or ACL set) successfully.
          headers:
            Location:
              description: URI of the newly created bucket.
              schema:
                type: string
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "409":
          description: Bucket already exists or is owned by someone else.
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    # ----- DELETE /{Bucket} -----------------------------------------------
    delete:
      operationId: DeleteBucket
      summary: Delete an empty bucket
      tags: [Bucket]
      responses:
        "204":
          description: Bucket deleted successfully.
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Bucket is not empty.
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    # ----- HEAD /{Bucket} -------------------------------------------------
    head:
      operationId: HeadBucket
      summary: Check whether a bucket exists and you have permission to access it
      tags: [Bucket]
      responses:
        "200":
          description: Bucket exists and is accessible.
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
            x-amz-bucket-region:
              description: Region where the bucket resides.
              schema:
                type: string
        "403":
          description: Access denied (no body for HEAD).
        "404":
          description: Bucket does not exist (no body for HEAD).
        "500":
          description: Internal server error (no body for HEAD).

    # ----- GET /{Bucket} --------------------------------------------------
    get:
      operationId: GetBucket
      summary: |
        Multiplex endpoint: ListObjectsV2 (list-type=2), ListObjects (legacy),
        GetBucketLocation (?location), GetBucketAcl (?acl), or
        ListMultipartUploads (?uploads).
      tags: [Bucket, Object, MultipartUpload]
      parameters:
        # -- discriminator query params --
        - name: list-type
          in: query
          description: Set to `2` to use ListObjectsV2.
          schema:
            type: string
            enum: ["2"]
        - name: location
          in: query
          description: When present (no value needed), returns the bucket location.
          allowEmptyValue: true
          schema:
            type: string
        - name: acl
          in: query
          description: When present (no value needed), returns the bucket ACL.
          allowEmptyValue: true
          schema:
            type: string
        - name: uploads
          in: query
          description: When present (no value needed), lists in-progress multipart uploads.
          allowEmptyValue: true
          schema:
            type: string
        # -- ListObjects / ListObjectsV2 common params --
        - name: prefix
          in: query
          description: Limits the response to keys beginning with this prefix.
          schema:
            type: string
        - name: delimiter
          in: query
          description: Character used to group keys (commonly `/`).
          schema:
            type: string
        - name: max-keys
          in: query
          description: Maximum number of keys to return (default 1000).
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 1000
        - name: encoding-type
          in: query
          description: Encode keys in the response (only `url` is supported).
          schema:
            type: string
            enum: [url]
        # -- ListObjectsV2 specific --
        - name: continuation-token
          in: query
          description: Token from a previous ListObjectsV2 response to continue listing.
          schema:
            type: string
        - name: start-after
          in: query
          description: Start listing after this key (ListObjectsV2 only).
          schema:
            type: string
        - name: fetch-owner
          in: query
          description: Set to true to include owner info in ListObjectsV2 results.
          schema:
            type: boolean
        # -- ListObjects (v1) specific --
        - name: marker
          in: query
          description: Key to start listing from (ListObjects v1 only).
          schema:
            type: string
        # -- ListMultipartUploads specific --
        - name: key-marker
          in: query
          description: Key at which to begin listing multipart uploads.
          schema:
            type: string
        - name: upload-id-marker
          in: query
          description: Upload ID at which to begin listing (used with key-marker).
          schema:
            type: string
        - name: max-uploads
          in: query
          description: Maximum number of multipart uploads to return.
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 1000
      responses:
        "200":
          description: |
            Response depends on the query parameters supplied:
            - No discriminator or `list-type=2`: object listing XML
            - `?location`: `LocationConstraint` XML
            - `?acl`: `AccessControlPolicy` XML
            - `?uploads`: `ListMultipartUploadsResult` XML
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
          content:
            application/xml:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ListBucketResult"
                  - $ref: "#/components/schemas/ListBucketResultV2"
                  - $ref: "#/components/schemas/LocationConstraint"
                  - $ref: "#/components/schemas/AccessControlPolicy"
                  - $ref: "#/components/schemas/ListMultipartUploadsResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    # ----- POST /{Bucket} -------------------------------------------------
    post:
      operationId: PostBucket
      summary: |
        Multiplex endpoint: DeleteObjects (?delete).
      tags: [Object]
      parameters:
        - name: delete
          in: query
          required: true
          description: Must be present to invoke multi-object delete.
          allowEmptyValue: true
          schema:
            type: string
        - name: Content-MD5
          in: header
          description: Base64-encoded 128-bit MD5 digest of the request body.
          required: true
          schema:
            type: string
        - name: x-amz-mfa
          in: header
          description: MFA device serial and token (for MFA-Delete enabled buckets).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: "#/components/schemas/DeleteObjectsRequest"
      responses:
        "200":
          description: Delete result (may contain both successful and failed deletions).
          headers:
            x-amz-request-id:
              $ref: "#/components/headers/x-amz-request-id"
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # =======================================================================
  # Object-level: /{Bucket}/{Key+}
  # =======================================================================
...
components:
...
  securitySchemes:
    sigv4:
      type: http
      scheme: bearer
      description: |
        AWS Signature Version 4 authentication. Clients must sign each
        request using their access key ID and secret access key, placing the
        signature in the Authorization header (or as query-string
        parameters for pre-signed URLs). The format is:

          Authorization: AWS4-HMAC-SHA256
            Credential=<access-key>/<date>/<region>/s3/aws4_request,
            SignedHeaders=<signed-headers>,
            Signature=<signature>

        BleepStore validates the signature against stored credentials.

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    BucketParam:
      name: Bucket
      in: path
      required: true
      description: Name of the S3 bucket.
      schema:
        type: string
        minLength: 3
        maxLength: 63
        pattern: "^[a-z0-9][a-z0-9.\\-]{1,61}[a-z0-9]$"
...
    AclQueryParam:
      name: acl
      in: query
      description: When present (no value needed), the operation targets the ACL sub-resource.
      allowEmptyValue: true
      schema:
        type: string
...
  headers:
    x-amz-request-id:
      description: Unique identifier assigned to the request by BleepStore.
      schema:
        type: string

  # -------------------------------------------------------------------------
  # Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: "400 -- The request was malformed or contained invalid parameters."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
    AccessDenied:
      description: "403 -- Access denied. The credentials are invalid or lack permission."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: "404 -- The specified bucket or object does not exist."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: "500 -- An internal server error occurred."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: "503 -- The service is temporarily unavailable. Retry with exponential backoff."
      headers:
        x-amz-request-id:
          $ref: "#/components/headers/x-amz-request-id"
      content:
        application/xml:
          schema:
            $ref: "#/components/schemas/Error"

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    CannedAcl:
      type: string
      enum:
        - private
        - public-read
        - public-read-write
        - authenticated-read
        - aws-exec-read
        - bucket-owner-read
        - bucket-owner-full-control
        - log-delivery-write
...
    Error:
      type: object
      xml:
        name: Error
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Code:
          type: string
          description: Machine-readable error code (e.g. NoSuchBucket, AccessDenied).
        Message:
          type: string
          description: Human-readable error description.
        Resource:
          type: string
          description: The resource associated with the error.
        RequestId:
          type: string
          description: Unique request identifier.
        HostId:
          type: string
          description: Extended request ID for debugging.
        BucketName:
          type: string
          description: Bucket name (present on some bucket-related errors).
        Key:
          type: string
          description: Object key (present on some object-related errors).
      required: [Code, Message, RequestId]

    # === Bucket list ======================================================
    ListAllMyBucketsResult:
      type: object
      xml:
        name: ListAllMyBucketsResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Owner:
          $ref: "#/components/schemas/Owner"
        Buckets:
          type: object
          properties:
            Bucket:
              type: array
              items:
                $ref: "#/components/schemas/BucketInfo"
              xml:
                wrapped: false
      required: [Owner, Buckets]
    BucketInfo:
      type: object
      xml:
        name: Bucket
      properties:
        Name:
          type: string
        CreationDate:
          type: string
          format: date-time
      required: [Name, CreationDate]

    # === Bucket configuration =============================================
    CreateBucketConfiguration:
      type: object
      xml:
        name: CreateBucketConfiguration
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        LocationConstraint:
          type: string
          description: Region where the bucket should be created.
    LocationConstraint:
      type: object
      xml:
        name: LocationConstraint
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        value:
          type: string
          description: The region name (element text content).
          xml:
            x-text: true

    # === ACL ==============================================================
    AccessControlPolicy:
      type: object
      xml:
        name: AccessControlPolicy
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Owner:
          $ref: "#/components/schemas/Owner"
        AccessControlList:
          type: object
          properties:
            Grant:
              type: array
              items:
                $ref: "#/components/schemas/Grant"
              xml:
                wrapped: false
      required: [Owner, AccessControlList]
    Owner:
      type: object
      properties:
        ID:
          type: string
          description: Canonical user ID.
        DisplayName:
          type: string
          description: Display name of the owner.
      required: [ID]
    Grant:
      type: object
      properties:
        Grantee:
          $ref: "#/components/schemas/Grantee"
        Permission:
          $ref: "#/components/schemas/Permission"
      required: [Grantee, Permission]
...
    ListBucketResult:
      type: object
      xml:
        name: ListBucketResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Name:
          type: string
          description: Bucket name.
        Prefix:
          type: string
        Marker:
          type: string
        NextMarker:
          type: string
          description: Present when IsTruncated is true and a delimiter was used.
        MaxKeys:
          type: integer
        Delimiter:
          type: string
        IsTruncated:
          type: boolean
        EncodingType:
          type: string
        Contents:
          type: array
          items:
            $ref: "#/components/schemas/ObjectEntry"
          xml:
            wrapped: false
        CommonPrefixes:
          type: array
          items:
            $ref: "#/components/schemas/CommonPrefix"
          xml:
            wrapped: false
      required: [Name, Prefix, MaxKeys, IsTruncated]

    # === ListObjectsV2 ====================================================
    ListBucketResultV2:
      type: object
      xml:
        name: ListBucketResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Name:
          type: string
        Prefix:
          type: string
        StartAfter:
          type: string
        ContinuationToken:
          type: string
        NextContinuationToken:
          type: string
          description: Present when IsTruncated is true.
        MaxKeys:
          type: integer
        Delimiter:
          type: string
        IsTruncated:
          type: boolean
        KeyCount:
          type: integer
          description: Number of keys returned in this response.
        EncodingType:
          type: string
        Contents:
          type: array
          items:
            $ref: "#/components/schemas/ObjectEntry"
          xml:
            wrapped: false
        CommonPrefixes:
          type: array
          items:
            $ref: "#/components/schemas/CommonPrefix"
          xml:
            wrapped: false
      required: [Name, Prefix, MaxKeys, KeyCount, IsTruncated]
    ObjectEntry:
      type: object
      xml:
        name: Contents
      properties:
        Key:
          type: string
        LastModified:
          type: string
          format: date-time
        ETag:
          type: string
        Size:
          type: integer
          format: int64
        StorageClass:
          $ref: "#/components/schemas/StorageClass"
        Owner:
          $ref: "#/components/schemas/Owner"
        ChecksumAlgorithm:
          type: string
          description: Checksum algorithm used (e.g., SHA256, CRC32).
      required: [Key, LastModified, ETag, Size, StorageClass]
    CommonPrefix:
      type: object
      xml:
        name: CommonPrefixes
      properties:
        Prefix:
          type: string
      required: [Prefix]

    # === CopyObject =======================================================
...
    DeleteObjectsRequest:
      type: object
      xml:
        name: Delete
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Quiet:
          type: boolean
          description: When true, the response only includes errors.
          default: false
        Object:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/ObjectIdentifier"
          xml:
            wrapped: false
      required: [Object]
    ObjectIdentifier:
      type: object
      xml:
        name: Object
      properties:
        Key:
          type: string
          description: Key of the object to delete.
        VersionId:
          type: string
          description: Version ID of the object to delete (optional).
      required: [Key]
    DeleteResult:
      type: object
      xml:
        name: DeleteResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Deleted:
          type: array
          items:
            $ref: "#/components/schemas/DeletedObject"
          xml:
            wrapped: false
        Error:
          type: array
          items:
            $ref: "#/components/schemas/DeleteError"
          xml:
            wrapped: false
...
    ListMultipartUploadsResult:
      type: object
      xml:
        name: ListMultipartUploadsResult
        namespace: http://s3.amazonaws.com/doc/2006-03-01/
      properties:
        Bucket:
          type: string
        KeyMarker:
          type: string
        UploadIdMarker:
          type: string
        NextKeyMarker:
          type: string
        NextUploadIdMarker:
          type: string
        MaxUploads:
          type: integer
        IsTruncated:
          type: boolean
        Prefix:
          type: string
        Delimiter:
          type: string
        EncodingType:
          type: string
        Upload:
          type: array
          items:
            $ref: "#/components/schemas/MultipartUploadEntry"
          xml:
            wrapped: false
        CommonPrefixes:
          type: array
          items:
            $ref: "#/components/schemas/CommonPrefix"
          xml:
            wrapped: false
      required: [Bucket, KeyMarker, UploadIdMarker, MaxUploads, IsTruncated]
    MultipartUploadEntry:
      type: object
      xml:
        name: Upload
      properties:
        Key:
          type: string
        UploadId:
          type: string
        Initiator:
          $ref: "#/components/schemas/Initiator"
        Owner:
          $ref: "#/components/schemas/Owner"
        StorageClass:
          $ref: "#/components/schemas/StorageClass"
        Initiated:
          type: string
          format: date-time
          description: Date and time the multipart upload was initiated.
      required: [Key, UploadId, Initiator, Owner, StorageClass, Initiated]
...